Fix : manage login state only in Header #83
